Move stack-related code to buildtools/stack/
